+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.17] - 2021-01-26
+### Added
+- Handle SET_PERMISSION events and update the "shared_users" field
+
+### Changed
+- Don't throw a RuntimeError on an unhandled event; only log a warning
+
 ## [1.9.16] - 2021-01-20
 ### Fixed
 - AMA index runner configuration bug fix